// src/enums/esstateProperties.enum.ts
export enum EPROPERTIES {
  PRICE = 'Price',
  ADDRESS = 'Address',
  MLS_NUMBER = 'MLS number',
  BEDROOMS_COUNT = 'Bedrooms count',
  BATHROOMS_COUNT = 'Bathrooms count',
  DESCRIPTION = 'Description',
  PROPERTY_TYPE = 'Property type',
  BUILDING_TYPE = 'Building type',
  SQUARE_FOOTAGE = 'Square footage',
  NEIGHBORHOOD_NAME = 'Neighborhood name',
  TITLE = 'Title',
  LAND_SIZE = 'Land size',
  BUILT_IN = 'Built in',
  ANNUAL_PROPERTY_TAXES = 'Annual property taxes',
  PARKING_TYPE = 'Parking type',
  TIME_ON_PLATFORM = 'Time on platform',
  INTERiOR_FEATURES = 'Interior features',
  BUILDING_FEATURES = 'Building features',
  ARCHITECTURAL_STYLE = 'Architectural style',
  SRTUCTURES = 'Structures',
  LINK = 'link',
}